Code refactor - moved encryption module to elekto/core - updated exception handlign in decrypt()
